refactor: share the UDF codec between the logical and physical plan codecs
`scan::codec` (logical) and `scan::physical_codec` (physical) each carried their
own copy of the encode/decode logic for the two scalar UDFs pg_search puts into
DataFusion plans, `SearchPredicateUDF` and `PgExprUdf`. The bodies were
identical apart from two things, one of which was a bug.

The copies had drifted on error handling. When a `SearchPredicateUDF` carries a
`plan_position` with no matching canonical segment IDs, the physical codec
returns `DataFusionError::Internal`, while the logical codec called `.expect()`
and panicked. Both run inside a Postgres backend, where a panic is considerably
worse than a plan-time error. The shared helper uses the error path, so the
logical codec no longer panics on that input.

The genuine difference is the fallback for a UDF that is not ours: the logical
codec must reject it, while the physical codec deliberately encodes nothing so
the expression travels by name and the decoding session resolves it from its
registry. The helpers therefore report "not mine" (`Ok(None)` / `Ok(false)`)
rather than deciding, and each codec keeps its own fallback.

`is_pg_search_udf`, which duplicated the same "is it ours" predicate a third
time, moves into the shared module so all three agree by construction.

Net 117 lines removed from the two codecs for 13 added, plus the shared module.
