+++ pg_search/src/scan/mod.rs
@@ -33,6 +33,7 @@ pub mod table_provider;
 pub mod tantivy_lookup_exec;
 #[cfg(any(test, feature = "pg_test"))]
 mod tests;
+mod udf_codec;
 pub mod visibility_ctid_resolver_rule;
 
 pub use batch_scanner::Scanner;
